Changed TalonSRXs so that they now function correctly
We added a follower clause to the two follower motors, and also
configured the encoders so that they give us the correct rpm and fps.
Finally, we made it so that the gear box switches to high gear
automatically at 6.4 feet per second, but it requires manual input to
change back to lower.